# app/server/src/openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Aureus Group - Take Home Assignment
  contact:
    name: Welson
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  description: |
    API Documentation for Job Portal
...
servers:
  - url: 'http://localhost:8888/api/v1'
    description: localhost server for testing purposes
paths:
...
  '/admin/login':
    post:
      tags:
        - Admin
      summary: Login as an admin into the system
      description: Login as an admin requiring email and password to be provided
      operationId: adminLogin
      security: 
        -  {}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: 
                email:
                  $ref: "#/components/schemas/Email"
                password:
                  $ref: "#/components/schemas/Password"
      responses:
        '200':
          description: Success with JWT token
          content: 
            application/jwt:
              schema: 
                type: string
        '400':
          description: Invalid Input
...
components:
  securitySchemes:
    AccessToken:
      description: JWT to identify an user
      bearerFormat: JWT
      scheme: bearer
      type: http
  schemas:
    Email:
      description: User's email address. This will be system's main way to identify user.
      type: string
      format: email
      example: example@example.com
    Password:
      description: User's Password
      type: string
      format: password
      minLength: 8
      writeOnly: true
      example: yourpassword
...